+++ src/main/java/com/microsoft/spring/data/gremlin/telemetry/TelemetryTracker.java
@@ -8,10 +8,17 @@ package com.microsoft.spring.data.gremlin.telemetry;
 import com.microsoft.applicationinsights.TelemetryClient;
 import org.springframework.lang.NonNull;
 
+import java.net.InetAddress;
+import java.net.NetworkInterface;
+import java.net.SocketException;
+import java.net.UnknownHostException;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.microsoft.applicationinsights.core.dependencies.apachecommons.codec.digest.DigestUtils.sha256Hex;
 import static com.microsoft.spring.data.gremlin.telemetry.TelemetryProperties.PROPERTY_INSTALLATION_ID;
+import static com.microsoft.spring.data.gremlin.telemetry.TelemetryProperties.PROPERTY_SERVICE_NAME;
 import static com.microsoft.spring.data.gremlin.telemetry.TelemetryProperties.PROPERTY_VERSION;
 
 public class TelemetryTracker {
@@ -20,21 +27,23 @@ public class TelemetryTracker {
 
     private static final String PROJECT_INFO = "spring-data-gremlin" + "/" + PROJECT_VERSION;
 
-    private TelemetryClient client;
+    private static final String SERVICE_NAME = "gremlin";
 
-    public TelemetryTracker() {
+    private static final String UNKNOWN_MAC = "unknown-Mac-Address";
+
+    private TelemetryClient client;
+    private boolean isAllowTelemetry;
+
+    public TelemetryTracker(boolean isAllowTelemetry) {
         this.client = new TelemetryClient();
+        this.isAllowTelemetry = isAllowTelemetry;
     }
 
-    public void trackEvent(@NonNull String name, Map<String, String> customProperties) {
-        final Map<String, String> properties;
-        final Map<String, String> defaultProperties = this.getDefaultProperties();
+    public void trackEvent(@NonNull String name) {
+        final Map<String, String> properties = this.getDefaultProperties();
 
-        if (customProperties == null) {
-            properties = defaultProperties;
-        } else {
-            defaultProperties.forEach(customProperties::putIfAbsent);
-            properties = customProperties;
+        if (isAllowTelemetry) {
+            properties.put(PROPERTY_INSTALLATION_ID, getHashMac());
         }
 
         client.trackEvent(name, properties, null);
@@ -45,8 +54,34 @@ public class TelemetryTracker {
         final Map<String, String> properties = new HashMap<>();
 
         properties.put(PROPERTY_VERSION, PROJECT_INFO);
-        properties.put(PROPERTY_INSTALLATION_ID, TelemetryUtils.getHashMac());
+        properties.put(PROPERTY_SERVICE_NAME, SERVICE_NAME);
 
         return properties;
+    }
+
+    private static String getMacAddress() {
+        final InetAddress ip;
+        final NetworkInterface network;
+        final byte[] macBytes;
+
+        try {
+            ip = InetAddress.getLocalHost();
+            network = NetworkInterface.getByInetAddress(ip);
+            macBytes = network.getHardwareAddress();
+        } catch (UnknownHostException | SocketException e) { // Omit
+            return UNKNOWN_MAC;
+        }
+
+        return Arrays.toString(macBytes);
+    }
+
+    private static String getHashMac() {
+        final String mac = getMacAddress();
+
+        if (mac.equals(UNKNOWN_MAC)) {
+            return UNKNOWN_MAC;
+        }
+
+        return sha256Hex(mac);
     }
 }
